requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.4
django-import-export==1.0.0
et-xmlfile==1.0.1
gunicorn==19.8.1
idna==2.6
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
ply==3.11
psycopg2==2.7.4
pyjsparser==2.5.2
PySocks==1.6.8
pytz==2018.4
PyYAML==3.12
requests==2.18.4
slimit==0.8.1
SocksiPy-branch==1.1
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
